+++ xmlserializer/XmlElement.h
@@ -61,7 +61,7 @@ public:
     // Navigation
     bool getChildElement(const std::string& strType, CXmlElement& childElement) const;
     bool getChildElement(const std::string& strType, const std::string& strNameAttribute, CXmlElement& childElement) const;
-    uint32_t getNbChildElements() const;
+    size_t getNbChildElements() const;
     bool getParentElement(CXmlElement& parentElement) const;
 
     // Setters
